Move non FrameworkBundle dependent fixtures to its own namespace
In preparation to introducing FrameworkBundle dependent fixtures.
