+++ tests/Symfony/TestCase/Fixtures/NoFrameworkBundle/AnotherTestKernel.php
@@ -1,16 +1,18 @@
 <?php
 declare(strict_types=1);
 
-namespace Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures;
+namespace Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\NoFrameworkBundle;
 
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\Kernel;
 use Zalas\Injector\PHPUnit\Symfony\Compiler\Discovery\ClassFinder;
 use Zalas\Injector\PHPUnit\Symfony\Compiler\Discovery\PropertyDiscovery;
 use Zalas\Injector\PHPUnit\Symfony\Compiler\ExposeServicesForTestsPass;
+use Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\Service1;
 
-class TestKernel extends Kernel
+class AnotherTestKernel extends Kernel
 {
     public function registerBundles()
     {
@@ -19,22 +21,18 @@ class TestKernel extends Kernel
 
     public function getCacheDir()
     {
-        return \sys_get_temp_dir().'/ZalasPHPUnitInjector/cache/'.$this->environment;
+        return \sys_get_temp_dir().'/ZalasPHPUnitInjector/NoFrameworkBundle/AnotherTestKernel/cache/'.$this->environment;
     }
 
     public function getLogDir()
     {
-        return \sys_get_temp_dir().'/ZalasPHPUnitInjector/logs';
+        return \sys_get_temp_dir().'/ZalasPHPUnitInjector/NoFrameworkBundle/AnotherTestKernel/logs';
     }
 
-    /**
-     * Loads the container configuration.
-     */
     public function registerContainerConfiguration(LoaderInterface $loader)
     {
         $loader->load(function (ContainerBuilder $container) use ($loader) {
             $container->register(Service1::class, Service1::class);
-            $container->register('foo.service2', Service2::class);
         });
     }
 
@@ -43,7 +41,7 @@ class TestKernel extends Kernel
         if ('test' === $this->getEnvironment()) {
             $container->addCompilerPass(
                 new ExposeServicesForTestsPass(
-                    new PropertyDiscovery(new ClassFinder(__DIR__ . '/../'))
+                    new PropertyDiscovery(new ClassFinder(__DIR__ . '/../../'))
                 )
             );
         }
