+++ tests/Symfony/TestCase/Fixtures/NoFrameworkBundle/TestKernel.php
@@ -1,16 +1,19 @@
 <?php
 declare(strict_types=1);
 
-namespace Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures;
+namespace Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\NoFrameworkBundle;
 
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\Kernel;
 use Zalas\Injector\PHPUnit\Symfony\Compiler\Discovery\ClassFinder;
 use Zalas\Injector\PHPUnit\Symfony\Compiler\Discovery\PropertyDiscovery;
 use Zalas\Injector\PHPUnit\Symfony\Compiler\ExposeServicesForTestsPass;
+use Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\Service1;
+use Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\Service2;
 
-class AnotherTestKernel extends Kernel
+class TestKernel extends Kernel
 {
     public function registerBundles()
     {
@@ -19,21 +22,19 @@ class AnotherTestKernel extends Kernel
 
     public function getCacheDir()
     {
-        return \sys_get_temp_dir().'/ZalasPHPUnitInjector2/cache/'.$this->environment;
+        return \sys_get_temp_dir().'/ZalasPHPUnitInjector/NoFrameworkBundle/TestKernel/cache/'.$this->environment;
     }
 
     public function getLogDir()
     {
-        return \sys_get_temp_dir().'/ZalasPHPUnitInjector2/logs';
+        return \sys_get_temp_dir().'/ZalasPHPUnitInjector/NoFrameworkBundle/TestKernel/logs';
     }
 
-    /**
-     * Loads the container configuration.
-     */
     public function registerContainerConfiguration(LoaderInterface $loader)
     {
         $loader->load(function (ContainerBuilder $container) use ($loader) {
             $container->register(Service1::class, Service1::class);
+            $container->register('foo.service2', Service2::class);
         });
     }
 
@@ -42,7 +43,7 @@ class AnotherTestKernel extends Kernel
         if ('test' === $this->getEnvironment()) {
             $container->addCompilerPass(
                 new ExposeServicesForTestsPass(
-                    new PropertyDiscovery(new ClassFinder(__DIR__ . '/../'))
+                    new PropertyDiscovery(new ClassFinder(__DIR__ . '/../../'))
                 )
             );
         }
