+++ tests/Symfony/TestCase/SymfonyContainerTest.php
@@ -27,7 +27,7 @@ class SymfonyContainerTest extends TestCase implements ServiceContainerTestCase
     private $service2;
 
     /**
-     * @env KERNEL_CLASS=Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\TestKernel
+     * @env KERNEL_CLASS=Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\NoFrameworkBundle\TestKernel
      */
     public function test_it_initializes_the_container_by_booting_the_symfony_kernel()
     {
@@ -41,7 +41,7 @@ class SymfonyContainerTest extends TestCase implements ServiceContainerTestCase
     }
 
     /**
-     * @env KERNEL_CLASS=Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\AnotherTestKernel
+     * @env KERNEL_CLASS=Zalas\Injector\PHPUnit\Tests\Symfony\TestCase\Fixtures\NoFrameworkBundle\AnotherTestKernel
      */
     public function test_it_ignores_missing_services_when_registering_the_service_locator()
     {
@@ -49,7 +49,7 @@ class SymfonyContainerTest extends TestCase implements ServiceContainerTestCase
 
         $this->assertInstanceOf(ServiceLocator::class, $container, 'Full container is not exposed.');
         $this->assertTrue($container->has(Service1::class), 'The private service is available in tests.');
-        $this->assertFalse($container->has('foo.service2'), 'The private service is available in tests.');
+        $this->assertFalse($container->has('foo.service2'), 'The private service is not available in tests.');
         $this->assertInstanceOf(Service1::class, $container->get(Service1::class));
     }
 }
